Begin minor refactor
